Move types to corresponding api module and move VM tests to own file
