+++ src/lib.rs
@@ -6,7 +6,7 @@ mod types;
 #[macro_use]
 mod macros;
 
-pub use api::{vm::Vm, Client, RestartError, RevertSnapshotError};
+pub use api::Client;
 pub use jsonrpsee_types::{Error as RpcError, JsonValue};
 pub use object_type::ObjectType;
-pub use types::{Snapshot, SnapshotId, Subscription, VmId};
+pub use types::Subscription;
